requirements.txt
Django>=1.3
djangorestframework>=2.3.14
mock==1.0.1
simplejson==3.6.0
